Fix: Correctly handle multiple entries with equal timestamp.
If two taxi entries for the same taxi at the same time appear there is
no meaningful value to compute. In this case, only the first tuple at
that time is considered.
Future tuples will have speed 0 (it is assumed that the taxi stays at
the position of the first tuple), the traveled distance will not change
and the average speed will also stay unaffected.

Added two tests to ensure this property.